requirements.txt
# Core Framework
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0

# AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0
sentence-transformers>=2.2.0

# Financial Data
yfinance>=0.2.18
alpha-vantage>=2.3.1
pandas-datareader>=0.10.0
fredapi>=0.5.0

# Portfolio Analytics
scipy>=1.10.0
scikit-learn>=1.3.0
cvxpy>=1.3.0
PyPortfolioOpt>=1.5.0
quantlib>=1.31

# Data Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.0.0

# NLP & Text Processing
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0
newspaper3k>=0.2.8

# API & Web
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Database & Storage
sqlalchemy>=2.0.0
sqlite3  # Built-in

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
tqdm>=4.65.0
loguru>=0.7.0

# Power BI Integration (Optional - for advanced features)
azure-storage-blob>=12.17.0
azure-servicebus>=7.11.0

# Additional ML Tools
optuna>=3.2.0  # Hyperparameter optimization
shap>=0.42.0   # Model explainability
